redis deployment setting
