+++ src/chat/src/repository.ts
@@ -1,16 +1,7 @@
-import mysql, { RowDataPacket } from "mysql2/promise";
-import dotenv from "dotenv";
+import { RowDataPacket } from "mysql2/promise";
 import { StoreMessageDataModel, CreateRoomModel } from "./model";
 import { generateChatroomID } from "./service";
-
-dotenv.config();
-
-const mysqlPool = mysql.createPool({
-  host: process.env.MYSQL_HOST,
-  user: process.env.MYSQL_USER,
-  password: process.env.MYSQL_PASSWORD,
-  database: process.env.MYSQL_DATABASE,
-});
+import mysqlPool from "./utils/mysql";
 
 export async function storeMessageRepository(data: StoreMessageDataModel[]) {
   const connection = await mysqlPool.getConnection();
