+++ src/chat/src/utils/mysql.ts
@@ -1,0 +1,15 @@
+import mysql from "mysql2/promise";
+import dotenv from "dotenv";
+import process from "node:process";
+
+dotenv.config();
+
+const mysqlPool = mysql.createPool({
+  host: process.env.MYSQL_HOST,
+  user: process.env.MYSQL_USER,
+  password: process.env.MYSQL_PASSWORD,
+  database: process.env.MYSQL_DATABASE,
+  port: Number(process.env.MYSQL_PORT),
+});
+
+export default mysqlPool;
